- merged in changes from experimental branch, to coincide with 5.1.2 release

+++ Common/Source/MapWindow2.cpp
@@ -445,9 +445,9 @@ rectObj MapWindow::CalculateScreenBounds(double scale) {
 
   } else {
 
-    float xmin, xmax, ymin, ymax;
+    double xmin, xmax, ymin, ymax;
     int x, y;
-    float X, Y;
+    double X, Y;
 
     x = MapRect.left;
     y = MapRect.top;
@@ -472,6 +472,7 @@ rectObj MapWindow::CalculateScreenBounds(double scale) {
     Screen2LatLon(x, y, X, Y);
     xmin = min(xmin, X); xmax = max(xmax, X);
     ymin = min(ymin, Y); ymax = max(ymax, Y);
+
 
     sb.minx = xmin;
     sb.maxx = xmax;
@@ -545,6 +546,7 @@ void MapWindow::CalculateScreenPositionsThermalSources() {
         DerivedDrawInfo.ThermalSources[i].Visible = false;
         continue;
       }
+
       double t = dh/DerivedDrawInfo.ThermalSources[i].LiftRate;
       double lat, lon;
       FindLatitudeLongitude(DerivedDrawInfo.ThermalSources[i].Latitude,
